Bugfix for the most bottom/right item being stuck. (#828)
Item was stuck when the boundary control was "true". Because the lastMouse position for item that is right and bottom, is at first clientX = 0, and clientY = 0. And this gives us the wrong direction we want to move item. Therefore item is not moved, and lastMouse position not updates, as the direction.length is 0.