Bugfix: lag when dragging a widget with boundary control enabled (#847)
* Bugfix lag when dragging a widget with boundary control enabled

There is a lag when enableBoundaryControl is true and we start dragging widget which touches the boundary of the grid. When gridster item and grid positions are same then dragging should be allowed. Removed = to fail the 4 if conditions in this scenario and allow dragging.

* Bugfix fix typo for changes made in #828

Co-authored-by: Yashwanth Ande <yande@opentext.com>
(cherry picked from commit 4a6e51d7a69dec74d54261003983237897f053f0)